+++ src/main/java/com/example/evm/repository/vehicle/VehicleRepository.java
@@ -3,4 +3,4 @@ package com.example.evm.repository.vehicle;
 import com.example.evm.entity.vehicle.Vehicle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleRepository extends JpaRepository<Vehicle, Integer> {}
+public interface VehicleRepository extends JpaRepository<Vehicle, Long> {}
