+++ src/main/java/com/example/evm/repository/vehicle/VehicleVariantRepository.java
@@ -3,5 +3,5 @@ package com.example.evm.repository.vehicle;
 import com.example.evm.entity.vehicle.VehicleVariant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleVariantRepository extends JpaRepository<VehicleVariant, Integer> {
+public interface VehicleVariantRepository extends JpaRepository<VehicleVariant, Long> {
 }
